requirements.txt
Flask
Pillow
tensorflow
numpy
